+++ src/WMTPlatformUtils.ts
@@ -14,54 +14,17 @@
 // and limitations under the License.
 //
 
-import { WMT_SDK_VERSION } from "./WMTSDKVersion"
 import { Platform } from "react-native"
-import { WMTLogger } from "./WMTLogger"
-import { PowerAuthEnvironmentInfo, PowerAuthUtils } from "react-native-powerauth-mobile-sdk"
+import { WMTDefaultUserAgent } from "./utils/WMTDefaultUserAgent"
 
 /* @internal */
 export class WMTPlatformUtils {
-
-    private static cachedEnvironmentInfo?: PowerAuthEnvironmentInfo
 
     static getPlatform():  "ios" | "android" {
         return Platform.OS == "ios" ? "ios" : "android"
     }
 
     static async getDefaultUserAgent(): Promise<string> {
-        const product = "MobileTokenJS"
-        const sdkVer = WMT_SDK_VERSION
-        const envInfo = await this.getEnvironmentInfo()
-        const appVer = envInfo.applicationVersion || "0.0"
-        const appId = envInfo.applicationIdentifier || "unknown"
-        const maker = envInfo.deviceManufacturer
-        const model = envInfo.deviceId
-        const os = envInfo.systemName
-        const osVer = envInfo.systemVersion
-        const userAgent = `${product}/${sdkVer} ${appId}/${appVer} (${maker}; ${os}/${osVer}; ${model}`
-        return userAgent
-    }
-
-    private static async getEnvironmentInfo(): Promise<PowerAuthEnvironmentInfo> {
-        try {
-            // If we have cached environment info, return it to avoid unnecessary calls.
-            // This expects that the environment info does not change during the app lifetime.
-            if (!this.cachedEnvironmentInfo) {
-                this.cachedEnvironmentInfo = await PowerAuthUtils.getEnvironmentInfo();
-            }
-            return this.cachedEnvironmentInfo
-        } catch (e) {
-            WMTLogger.error(`Failed to get environment info: ${e}`)
-            // In case of error, we return a default object with "unknown" values.
-            return {
-                systemName: "unknown",
-                systemVersion: "0.0",
-                applicationVersion: "0.0",
-                applicationIdentifier: "unknown",
-                deviceManufacturer: "unknown",
-                deviceId: "unknown",
-                sdkVersion: "0.0"
-            }
-        }
+        return WMTDefaultUserAgent.get()
     }
 }