+++ docs/Changelog.md
@@ -1,14 +1,16 @@
 # Changelog
 
-## 2.0.0
+## 2.0.0 (TBA)
 
-- Minimal required version for `react-native-powerauth-mobile-sdk` raised to `4.1.0`
+- Minimal required version for `PowerAuth Mobile JS SDK` raised to `4.1.0`
+- Removed `react-native-device-info` dependency for react-native
+- `WMTPlatformUtils.getDefaultUserAgent` now return promise (`Promise<string>`)
 
 ## 1.1.0 (TBA)
 
 - Removed `react-native-device-info` dependency for react-native
 - `WMTPlatformUtils.getDefaultUserAgent` now return promise (`Promise<string>`)
-- Minimal required version for `react-native-powerauth-mobile-sdk` raised to `3.1.0`
+- Minimal required version for `PowerAuth Mobile JS SDK` raised to `3.2.0`
 
 ## 1.0.0 (2/2025)